week_12/day_01/introToJava/homework/week_12_day_1_homework/src/main/java/Printer.java
public class Printer {

    private int paper;
    private int toner;

    public Printer(int paper, int toner) {
        this.paper = paper;
        this.toner = toner;
    }

    public void print(int pages, int copies){
        if(paper >= pages * copies){
            paper -= pages * copies;
            toner -= pages * copies;
        }
    }

    public int getPaper() {
        return paper;
    }

    public void setPaper(int paper) {
        this.paper = paper;
    }

    public int getToner() {
        return toner;
    }

    public void setToner(int toner) {
        this.toner = toner;
    }
}
